revert conversion of ean and upc readers temporarily
- i have not yet been able to get ean_reader to work properly when
  in typescript.  going to convert them to class-inheritance javascript
  first, and see if that works, then go to typescript from there.
